Fix "expected u16, found i32" compilation error (#135)
* Fix "expected u16, found i32" compilation error

* Update Cargo.lock

* Revert "Update Cargo.lock"

This reverts commit 3f58123da3b9a68d2ca4545d3e2e211de341f0df.

* Update cargo.lock (revert adding main.rs)

+++ src/main_loop.rs
@@ -144,7 +144,7 @@ impl Future for MainLoopState {
                     ConnectConfig {
                         name: self.spotifyd_state.device_name.clone(),
                         device_type: DeviceType::default(),
-                        volume: i32::from(mixer.volume()),
+                        volume: u16::from(mixer.volume()),
                         linear_volume: self.linear_volume,
                     },
                     session.clone(),
